+++ test/unit/client.test.ts
@@ -8,13 +8,13 @@ import { AxiosError, AxiosHeaders } from 'axios';
 import SaslError from '../../src/types/sasl-error';
 import { TimeoutError } from '../../src/types/timeout-error';
 import OfflineError from '../../src/types/offline-error';
-import { reject } from 'lodash';
 import EventEmitter from 'events';
 import { NamedAgent } from '../../src/types/named-agent';
 import { flushPromises } from '../helpers/testing-utils';
 import { SCConnectionData, StreamingClientErrorTypes, StreamingClientError } from '../../src';
 import { Ping } from '../../src/ping';
 import { ServerMonitor } from '../../src/server-monitor';
+import UserCancelledError from '../../src/types/user-cancelled-error';
 
 jest.mock('genesys-cloud-client-logger');
 jest.mock('../../src/ping');
@@ -270,6 +270,14 @@ describe('connect', () => {
     backoffRetrySpy = client['backoffConnectRetryHandler'] = jest.fn();
   });
 
+  it('should set cancelConnectionAttempt to false', async () => {
+    connectionAttemptSpy.mockResolvedValue(null);
+    client['cancelConnectionAttempt'] = true;
+
+    await client.connect();
+    expect(client['cancelConnectionAttempt']).toBeFalsy();
+  });
+
   it('should do nothing if already connecting', async () => {
     client.connecting = true;
     await client.connect();
@@ -335,6 +343,25 @@ describe('connect', () => {
     } catch (err) {
       expect(err).toBeInstanceOf(StreamingClientError);
       expect(err['type']).toBe(StreamingClientErrorTypes.generic);
+      expect(err['details']).toBe(error);
+    }
+    expect(connectionAttemptSpy).toHaveBeenCalledTimes(1);
+    expect.assertions(4);
+  });
+
+  it('should throw a user_cancelled error if the connection attempt was cancelled regardless of the thrown error', async () => {
+    const error = new SaslError('incorrect-encoding', 'channelId', 'instanceId');
+    connectionAttemptSpy.mockRejectedValue(error);
+    connectionAttemptSpy.mockImplementation(() => {
+      client['cancelConnectionAttempt'] = true;
+      throw error;
+    });
+
+    try {
+      await client.connect({ keepTryingOnFailure: false });
+    } catch (err) {
+      expect(err).toBeInstanceOf(StreamingClientError);
+      expect(err['type']).toBe(StreamingClientErrorTypes.userCancelled);
       expect(err['details']).toBe(error);
     }
     expect(connectionAttemptSpy).toHaveBeenCalledTimes(1);
@@ -521,22 +548,15 @@ describe('disconnect', () => {
   beforeEach(() => {
     client = new Client(getDefaultOptions());
     client.hardReconnectRequired = false;
+    client['autoReconnect'] = true;
   });
 
-  it('should do nothing if no stanza instance', async () => {
-    client.activeStanzaInstance = undefined;
-
-    const spy = client.http.stopAllRetries = jest.fn();
-
-    await client.disconnect();
-
-    expect(spy).not.toHaveBeenCalled();
-  });
-
-  it('should resolve when a disconnected event is received', async () => {
+  it('should stop HTTP retries and resolve when there is a current stanza instance', async () => {
     let isResolved = false;
 
-    const stanza = client.activeStanzaInstance = new EventEmitter() as any;
+    client.http.stopAllRetries = jest.fn();
+    const stanza = new EventEmitter() as any;
+    client['connectionManager'].currentStanzaInstance = stanza;
 
     let resolve;
     stanza.disconnect = jest.fn().mockImplementation(() => {
@@ -544,11 +564,11 @@ describe('disconnect', () => {
         resolve = r;
       });
     });
-    client.http.stopAllRetries = jest.fn();
 
-    const promise = client.disconnect().then(() => isResolved = true);
+    client.disconnect().then(() => isResolved = true);
     await flushPromises();
 
+    expect(client.http.stopAllRetries).toHaveBeenCalled();
     expect(stanza.disconnect).toHaveBeenCalled();
     expect(isResolved).toBeFalsy();
 
@@ -557,6 +577,23 @@ describe('disconnect', () => {
 
     expect(isResolved).toBeTruthy();
     expect(client.hardReconnectRequired).toBeTruthy();
+    expect(client['autoReconnect']).toBeFalsy();
+    expect(client['cancelConnectionAttempt']).toBeTruthy();
+  });
+
+  it('should stop HTTP retries and resolve when there is no current stanza instance', async () => {
+    let isResolved = false;
+
+    client.http.stopAllRetries = jest.fn();
+    client['connectionManager'].currentStanzaInstance = undefined;
+
+    await client.disconnect().then(() => isResolved = true);
+
+    expect(client.http.stopAllRetries).toHaveBeenCalled();
+    expect(isResolved).toBeTruthy();
+    expect(client.hardReconnectRequired).toBeTruthy();
+    expect(client['autoReconnect']).toBeFalsy();
+    expect(client['cancelConnectionAttempt']).toBeTruthy();
   });
 });
 
@@ -570,6 +607,21 @@ describe('backoffConnectRetryHandler', () => {
     });
 
     errorSpy = jest.spyOn(client.logger, 'error');
+  });
+
+  it('should return false if cancelConnectionAttempt is true', async () => {
+    const error = new UserCancelledError('user cancelled');
+    client['cancelConnectionAttempt'] = true;
+    const result = await client['backoffConnectRetryHandler']({ maxConnectionAttempts: 10 }, error, 1);
+    expect(result).toBeFalsy();
+  });
+
+  it('should return false if cancelConnectionAttempt is true even if called with another error', async () => {
+    const error = new TimeoutError('fake timeout');
+    client['cancelConnectionAttempt'] = true;
+
+    const result = await client['backoffConnectRetryHandler']({ maxConnectionAttempts: 2 }, error, 1);
+    expect(result).toBeFalsy();
   });
 
   it('should return false if not keepTryingOnFailure', async () => {
@@ -796,6 +848,14 @@ describe('makeConnectionAttempt', () => {
     pingerMock.mockClear();
   });
 
+  it('should not attempt if the connection attempt has been cancelled', async () => {
+    client['cancelConnectionAttempt'] = true;
+
+    await expect(client['makeConnectionAttempt']()).rejects.toThrow(UserCancelledError);
+    expect(prepareSpy).not.toHaveBeenCalled();
+    expect(getConnectionSpy).not.toHaveBeenCalled();
+  });
+
   it('should not attempt if offline', async () => {
     const spy = jest.spyOn(navigator, 'onLine', 'get').mockReturnValue(false)
 
@@ -804,6 +864,15 @@ describe('makeConnectionAttempt', () => {
     expect(getConnectionSpy).not.toHaveBeenCalled();
 
     spy.mockRestore();
+  });
+
+  it('should not set up a stanza instance if the connection attempt is cancelled after preparing to connect', async () => {
+    prepareSpy.mockImplementation(() => client['cancelConnectionAttempt'] = true);
+
+    await expect(client['makeConnectionAttempt']()).rejects.toThrow(UserCancelledError);
+
+    expect(prepareSpy).toHaveBeenCalled();
+    expect(getConnectionSpy).not.toHaveBeenCalled();
   });
 
   it('should set up stanzaInstance and emit connected event', async () => {
@@ -1063,7 +1132,7 @@ describe('prepareForConnect', () => {
 
     setConfigSpy = client['connectionManager'].setConfig = jest.fn();
     httpSpy = client.http.requestApi = jest.fn().mockImplementation((path) => {
-      const promise = new Promise(resolve => {
+      const promise = new Promise((resolve, reject) => {
         if (path === 'users/me') {
           return resolve({ data: { chat: { jabberId: 'myRequestedJid' } } });
         } else if (path.startsWith('notifications/channels')) {